requirements.txt
paho-mqtt
PyYAML
hvac
influxdb
